Version 2.0.0.5, with restructured serialization code and tiered logging.
git-svn-id: https://openbiometrics.svn.sourceforge.net/svnroot/openbiometrics/OpenEBTS@181 53fe7bc9-f1be-4a37-98cb-3b35ff32843c

+++ OpenEBTS/Java/OpenEBTSJNIHelpers.c
@@ -94,7 +94,7 @@ jstring JNINewString(JNIEnv *env, const TCHAR *sz)
 
 #ifdef UNICODE
 
-	int 	nLen = _tcslen(sz);
+	int 	nLen = (int)_tcslen(sz);
 
 #ifdef WIN32
 	// UNICODE/WIN32: NewString takes 16-bit wide chars so we're good to go
